pyproject.toml
[tool.poetry]
name = "misc"
version = "0.1.0"
description = "Misc tools"
authors = ["Dmitry Pyzhov <lux.place@gmail.com>"]
readme = "README.md"
packages = [{include = "misc"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
